+++ lib/formatter/data_formatter.dart
@@ -1,10 +1,9 @@
 import 'package:brasil_fields/formatter/max_length_input_formatter.dart';
 import 'package:flutter/services.dart';
 
-/// Formata o valor do campo com a mascara de CPF ( XXX.XXX.XXX-XX )
-class CpfInputFormatter extends TextInputFormatter
-    with MaxLengthInputFormatter {
-  final int maxLength = 11;
+/// Formata o valor do campo com a mascara de data ( 01/01/1900 )
+class DataFormatter extends TextInputFormatter with MaxLengthInputFormatter {
+  final int maxLength = 8;
   @override
   TextEditingValue formatEditUpdate(
       TextEditingValue oldValue, TextEditingValue newValue) {
@@ -14,18 +13,15 @@ class CpfInputFormatter extends TextInputFormatter
     int usedSubstringIndex = 0;
     final StringBuffer newText = StringBuffer();
 
-    if (newTextLength >= 4) {
-      newText.write(newValue.text.substring(0, usedSubstringIndex = 3) + '.');
-      if (newValue.selection.end >= 3) selectionIndex++;
+    if (newTextLength >= 3) {
+      newText.write(newValue.text.substring(0, usedSubstringIndex = 2) + '/');
+      if (newValue.selection.end >= 2) selectionIndex++;
     }
-    if (newTextLength >= 7) {
-      newText.write(newValue.text.substring(3, usedSubstringIndex = 6) + '.');
-      if (newValue.selection.end >= 6) selectionIndex++;
+    if (newTextLength >= 5) {
+      newText.write(newValue.text.substring(4, usedSubstringIndex = 5) + '/');
+      if (newValue.selection.end >= 5) selectionIndex++;
     }
-    if (newTextLength >= 10) {
-      newText.write(newValue.text.substring(6, usedSubstringIndex = 9) + '-');
-      if (newValue.selection.end >= 9) selectionIndex++;
-    }
+
     if (newTextLength >= usedSubstringIndex)
       newText.write(newValue.text.substring(usedSubstringIndex));
     if (maxLength != null &&